+++ handlers/__init__.py
@@ -1,7 +1,15 @@
 from handler import Handler
-from singup import Singup
+from signup import Signup
 from welcome import Welcome
 from login import Login
 from logout import Logout
 from newpost import NewPost
-
+from singlepost import SinglePost
+from edit import Edit
+from delete import Delete
+from likes import Likes
+from comment import Comment
+from editcomment import EditComment
+from deletecomment import DeleteComment
+from mainpage import MainPage
+from errorhandler import ErrorHandler
